removed overlycomplicated byte[] optimisation to int[] and changed messaging formats

+++ Communicator/src/main/java/com/george200150/bsc/model/Bitmap.java
@@ -18,7 +18,7 @@ public class Bitmap {
     private int width;
 
     @JsonProperty("pixels")
-    private byte[] pixels;
+    private int[] pixels;
 //    private List<Pixel> pixels;
 
 }
